feat: solve modules task
