+++ src/modules/cjsToEsm.mjs
@@ -1,0 +1,37 @@
+import * as path from 'path';
+import { fileURLToPath } from 'url';
+import { release, version } from 'os';
+import { createServer as createServerHttp } from 'http';
+import { readFile } from 'fs/promises';
+import './files/c.js';
+
+const objA = JSON.parse(await readFile(new URL('./files/a.json', import.meta.url)));
+const objB = JSON.parse(await readFile(new URL('./files/b.json', import.meta.url)));
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+const random = Math.random();
+
+let unknownObject;
+
+if (random > 0.5) {
+    unknownObject = {...objA};
+} else {
+    unknownObject = {...objB};
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+const createMyServer = createServerHttp((_, res) => {
+    res.end('Request accepted');
+});
+
+export {
+    unknownObject,
+    createMyServer,
+};
+
